remove extension, parameter access in standalone models
