+++ src/standalone/Bucket/Bucket.jl
@@ -1,4 +1,5 @@
 module Bucket
+import ClimaParams as CP
 using DocStringExtensions
 using Thermodynamics
 using Dates
@@ -241,6 +242,85 @@ Base.@kwdef struct BucketModelParameters{
     earth_param_set::PSE
 end
 
+##For interfacing with ClimaParams
+"""
+    BucketModelParameters(
+        ::Type{FT};
+        albedo,
+        z_0m,
+        z_0b,
+        τc,
+        kwargs...,
+    )
+
+    BucketModelParameters(
+        toml_dict::CP.AbstractTOMLDict;
+        albedo,
+        z_0m,
+        z_0b,
+        τc,
+        kwargs...,
+    )
+
+BucketModelParameters has a float-type and a toml-dict based constructor.
+Keyword arguments can be used to manually override any of the values in the struct.
+```julia
+BucketModelParameters(Float64; albedo, z_0m, z_0b, τc)
+BucketModelParameters(toml_dict; albedo, z_0m, z_0b, τc)
+```
+"""
+function BucketModelParameters(
+    ::Type{FT};
+    albedo,
+    z_0m,
+    z_0b,
+    τc,
+    kwargs...,
+) where {FT <: AbstractFloat}
+    return BucketModelParameters(
+        CP.create_toml_dict(FT);
+        albedo,
+        z_0m,
+        z_0b,
+        τc,
+        kwargs...,
+    )
+end
+
+function BucketModelParameters(
+    toml_dict::CP.AbstractTOMLDict;
+    albedo,
+    z_0m,
+    z_0b,
+    τc,
+    kwargs...,
+)
+
+    name_map = (;
+        :soil_conductivity => :κ_soil,
+        :soil_heat_capacity => :ρc_soil,
+        :critical_snow_water_equivalent => :σS_c,
+        :land_bucket_capacity => :W_f,
+        :critical_snow_fraction => :f_snow,
+        :bucket_capacity_fraction => :f_bucket,
+        :bucket_beta_decay_exponent => :p,
+    )
+    parameters = CP.get_parameter_values(toml_dict, name_map, "Land")
+
+    AAM = typeof(albedo)
+    earth_param_set = LP.LandParameters(toml_dict)
+    PSE = typeof(earth_param_set)
+    FT = CP.float_type(toml_dict)
+    BucketModelParameters{FT, AAM, PSE}(;
+        albedo,
+        z_0m,
+        z_0b,
+        τc,
+        earth_param_set,
+        parameters...,
+        kwargs...,
+    )
+end
 """
 
     struct BucketModel{
