+++ src/standalone/Snow/Snow.jl
@@ -1,5 +1,5 @@
 module Snow
-
+import ClimaParams as CP
 using DocStringExtensions
 import ...Parameters as LP
 using ClimaCore
@@ -329,6 +329,66 @@ function SnowParameters{FT}(
         ΔS,
         scf,
         earth_param_set,
+    )
+end
+
+## For interfacing with ClimaParams
+"""
+    function SnowParameters(
+        FT,
+        Δt;
+        kwargs...  # For individual parameter overrides
+    )
+
+    function SnowParameters(
+        toml_dict::CP.AbstractTOMLDict,
+        Δt;
+        kwargs...  # For individual parameter overrides
+    )
+
+Constructors for the SnowParameters struct. Two variants:
+1. Pass in the float-type and retrieve parameter values from the default TOML dict.
+2. Pass in a TOML dictionary to retrieve parameter values. Possible calls:
+```julia
+Δt = 450.0
+ClimaLand.Canopy.SnowParameters(Float64, Δt) # use the default values for all parameters
+# Kwarg overrides
+ClimaLand.Canopy.SnowParameters(Float64, Δt; ϵ_snow = 0.99)
+# TOML Dictionary:
+import ClimaParams as CP
+toml_dict = CP.create_toml_dict(Float32);
+ClimaLand.Canopy.SnowParameters(toml_dict, Δt; ϵ_snow = Float32(0.99), Ksat = Float32(1e-4))
+```
+"""
+SnowParameters(::Type{FT}, Δt; kwargs...) where {FT <: AbstractFloat} =
+    SnowParameters(CP.create_toml_dict(FT), Δt; kwargs...)
+
+function SnowParameters(toml_dict::CP.AbstractTOMLDict, Δt; kwargs...)
+    name_map = (;
+        :snow_momentum_roughness_length => :z_0m,
+        :snow_scalar_roughness_length => :z_0b,
+        :thermal_conductivity_of_water_ice => :κ_ice,
+        :snow_emissivity => :ϵ_snow,
+        :holding_capacity_of_water_in_snow => :θ_r,
+        :wet_snow_hydraulic_conductivity => :Ksat,
+    )
+
+    parameters = CP.get_parameter_values(toml_dict, name_map, "Land")
+    name_map2 = (; :snow_density => :ρ_snow, :snow_albedo => :α_snow)
+    parameters2 = CP.get_parameter_values(toml_dict, name_map2, "Land")
+
+    density = MinimumDensityModel(parameters2.ρ_snow)
+    α_snow = ConstantAlbedoModel(parameters2.α_snow)
+    FT = CP.float_type(toml_dict)
+    earth_param_set = LP.LandParameters(toml_dict)
+    PSE = typeof(earth_param_set)
+    return SnowParameters{FT}(
+        Δt;
+        earth_param_set,
+        parameters...,
+        density,
+        α_snow,
+        kwargs...,
     )
 end
