+++ src/standalone/Soil/Biogeochemistry/Biogeochemistry.jl
@@ -1,5 +1,6 @@
 module Biogeochemistry
 using ClimaLand
+import ClimaParams as CP
 using DocStringExtensions
 using ClimaCore
 import ...Parameters as LP
@@ -64,6 +65,41 @@ Base.@kwdef struct SoilCO2ModelParameters{FT <: AbstractFloat, PSE}
     p_sx::FT
     "Physical constants used Clima-wide"
     earth_param_set::PSE
+end
+
+## For interfacting with ClimaParams
+
+
+"""
+    SoilCO2ModelParameters(FT; kwargs...)
+    SoilCO2ModelParameters(toml_dict; kwargs...)
+
+SoilCO2ModelParameters has two constructors: float-type and toml dict based.
+Keywords arguments can be used to directly override any parameters.
+"""
+SoilCO2ModelParameters(::Type{FT}; kwargs...) where {FT <: AbstractFloat} =
+    SoilCO2ModelParameters(CP.create_toml_dict(FT); kwargs...)
+
+function SoilCO2ModelParameters(toml_dict::CP.AbstractTOMLDict; kwargs...)
+    name_map = (;
+        :CO2_diffusion_coefficient => :D_ref,
+        :soil_C_substrate_diffusivity => :D_liq,
+        :soilCO2_pre_expontential_factor => :α_sx,
+        :soilCO2_activation_energy => :Ea_sx,
+        :michaelis_constant => :kM_sx,
+        :O2_michaelis_constant => :kM_o2,
+        :O2_volume_fraction => :O2_a,
+        :oxygen_diffusion_coefficient => :D_oa,
+        :soluble_soil_carbon_fraction => :p_sx,
+    )
+    parameters = CP.get_parameter_values(toml_dict, name_map, "Land")
+    FT = CP.float_type(toml_dict)
+    earth_param_set = LP.LandParameters(toml_dict)
+    return SoilCO2ModelParameters{FT, typeof(earth_param_set)}(;
+        earth_param_set,
+        parameters...,
+        kwargs...,
+    )
 end
 
 """
