+++ src/standalone/Soil/Soil.jl
@@ -65,6 +65,7 @@ import ClimaCore.MatrixFields: @name, ⋅
 import ..Parameters as LP
 import ClimaCore: Fields, Operators, Geometry, Spaces
 using Thermodynamics
+import ClimaParams as CP
 using SurfaceFluxes
 using StaticArrays
 import SurfaceFluxes.Parameters as SFP
@@ -195,14 +196,17 @@ using .Biogeochemistry
 """
     EnergyHydrology(FT, domain, forcing, earth_param_set;
                          prognostic_land_components = (:soil),
-                         albedo = Soil.CLMTwoBandSoilAlbedo{FT}(; clm_soil_albedo_parameters(domain.space.surface)...),
-                         runoff =  ClimaLand.Soil.Runoff.TOPMODELRunoff{FT}(f_over = FT(3.28),
+                         albedo = CLMTwoBandSoilAlbedo{FT}(; clm_soil_albedo_parameters(domain.space.surface)...),
+                         runoff =  Runoff.TOPMODELRunoff{FT}(f_over = FT(3.28),
                                                                             R_sb = FT(1.484e-4 / 1000),
                                                                             f_max = topmodel_fmax(domain.space.surface,FT),
                                                                             ),
                          retention_parameters = soil_vangenuchten_parameters(domain.space.subsurface, FT),
                          composition_parameters = soil_composition_parameters(domain.space.subsurface, FT),
                          S_s = ClimaCore.Fields.zeros(domain.space.subsurface) .+ 1e-3,
+                         z_0m = LP.get_default_parameter(FT, :soil_momentum_roughness_length),
+                         z_0b = LP.get_default_parameter(FT, :soil_scalar_roughness_length),
+                         emissivity = LP.get_default_parameter(FT, :soil_bare_soil),
                          additional_sources = (),
                          )
 
@@ -218,7 +222,11 @@ The runoff and albedo parameterizations are also provided and can be changed via
 additional sources may be required in your model if the soil model will be composed with other 
 component models.
 
-TODO: Move scalar parameters to ClimaParams and obtain from earth_param_set, possibly use types in retention and composition arguments.
+Roughness lengths and soil emissivity are currently treated as constants; these can be passed in as Floats
+by kwarg; otherwise the default values are used.
+
+TODO: Move runoff scalar parameters to ClimaParams, possibly use types in retention, composition,
+roughness, and emissivity.
 """
 function EnergyHydrology(
     FT,
@@ -243,6 +251,9 @@ function EnergyHydrology(
         FT,
     ),
     S_s = ClimaCore.Fields.zeros(domain.space.subsurface) .+ 1e-3,
+    z_0m = LP.get_default_parameter(FT, :soil_momentum_roughness_length),
+    z_0b = LP.get_default_parameter(FT, :soil_scalar_roughness_length),
+    emissivity = LP.get_default_parameter(FT, :soil_bare_soil),
     additional_sources = (),
 )
     top_bc = AtmosDrivenFluxBC(
@@ -261,6 +272,9 @@ function EnergyHydrology(
         composition_parameters...,
         albedo,
         S_s,
+        z_0m,
+        z_b,
+        emissivity,
     )
     return EnergyHydrology{FT}(;
         parameters,
