+++ src/standalone/Soil/energy_hydrology.jl
@@ -75,6 +75,167 @@ end
 
 Base.broadcastable(ps::EnergyHydrologyParameters) = tuple(ps)
 
+
+"""
+    EnergyHydrologyParameters(
+        ::Type{FT};
+        ν,
+        ν_ss_om,
+        ν_ss_quartz,
+        ν_ss_gravel,
+        hydrology_cm,
+        K_sat,
+        S_s,
+        θ_r,
+        albedo = Soil.ConstantTwoBandSoilAlbedo{FT}(),
+        kwargs...,)
+
+    EnergyHydrologyParameters(
+        toml_dict;
+        ν,
+        ν_ss_om,
+        ν_ss_quartz,
+        ν_ss_gravel,
+        hydrology_cm,
+        K_sat,
+        S_s,
+        θ_r,
+        albedo = Soil.ConstantTwoBandSoilAlbedo{FT}(),
+        kwargs...,)
+
+EnergyHydrologyParameters has two constructors: float-type and toml dict based.
+Additional parameters must be added manually: ν, ν_ss_om, ν_ss_quartz, ν_ss_gravel, hydrology_cm, K_sat, S_s, and θ_r
+All parameters can be manually overriden via keyword arguments. Note, however,
+that certain parameters must have the same type (e.g, if a field is
+supplied for porosity, it must be supplied for all other parameters
+defined in the interior of the domain). Some parameters are defined only
+on the surface of the domain (e.g albedo), while other are defined everywhere
+(e.g. porosity). These are indicated with types `F` and `SF`. If both dry/wet albedos
+and general albedos are given as keywords, the dry/wet albedos will override the general
+albedos.
+
+Please see the EnergyHydrologyParameters documentation for a complete list.
+"""
+function EnergyHydrologyParameters(
+    ::Type{FT};
+    ν,
+    ν_ss_om,
+    ν_ss_quartz,
+    ν_ss_gravel,
+    hydrology_cm,
+    K_sat,
+    S_s,
+    θ_r,
+    albedo = Soil.ConstantTwoBandSoilAlbedo{FT}(),
+    kwargs...,
+) where {FT <: AbstractFloat}
+    return EnergyHydrologyParameters(
+        CP.create_toml_dict(FT);
+        ν,
+        ν_ss_om,
+        ν_ss_quartz,
+        ν_ss_gravel,
+        hydrology_cm,
+        K_sat,
+        S_s,
+        θ_r,
+        albedo,
+        kwargs...,
+    )
+end
+
+function EnergyHydrologyParameters(
+    toml_dict::CP.AbstractTOMLDict;
+    ν::F,
+    ν_ss_om::F,
+    ν_ss_quartz::F,
+    ν_ss_gravel::F,
+    hydrology_cm::C,
+    K_sat::F,
+    S_s::F,
+    θ_r::F,
+    albedo = Soil.ConstantTwoBandSoilAlbedo{FT}(),
+    kwargs...,
+) where {F <: Union{<:AbstractFloat, ClimaCore.Fields.Field}, C}
+    earth_param_set = LP.LandParameters(toml_dict)
+
+    # Obtain parameters needed to calculate the derived parameters
+    derived_param_name_map = (;
+        :thermal_conductivity_of_quartz => :κ_quartz,
+        :thermal_conductivity_of_soil_minerals => :κ_minerals,
+        :thermal_conductivity_of_water_ice => :κ_ice,
+        :thermal_conductivity_of_liquid_water => :κ_liq,
+        :thermal_conductivity_of_organic_matter => :κ_om,
+        :particle_density_quartz => :ρp_quartz,
+        :particle_density_minerals => :ρp_minerals,
+        :particle_density_organic_matter => :ρp_om,
+        :vol_heat_capacity_quartz => :ρc_quartz,
+        :vol_heat_capacity_organic_matter => :ρc_om,
+        :vol_heat_capacity_minerals => :ρc_minerals,
+    )
+    p = CP.get_parameter_values(toml_dict, derived_param_name_map, "Land")
+    # Particle density of the soil - per unit soil solids
+    # Denoted ρ_ds in the Clima Design Docs (Equation 2.3)
+    # where ν_ss_i = ν_i/(1-ν)
+    ρp = @. (
+        ν_ss_om * p.ρp_om +
+        ν_ss_quartz * p.ρp_quartz +
+        ν_ss_gravel * p.ρp_minerals +
+        (1 - ν_ss_om - ν_ss_quartz - ν_ss_gravel) * p.ρp_minerals
+    )
+    # Volumetric heat capacity of soil solids - per unit volume soil solids
+    # This is Equation 2.6a/2.10 in the Clima Design Docs
+    # where ν_ss_i = ν_i/(1-ν)
+    ρc_ss = @. (
+        ν_ss_om * p.ρc_om +
+        ν_ss_quartz * p.ρc_quartz +
+        ν_ss_gravel * p.ρc_minerals +
+        (1 - ν_ss_om - ν_ss_quartz - ν_ss_gravel) * p.ρc_minerals
+    )
+    # Volumetric heat capacity of dry soil - per unit volume of soil
+    # This is denoted č_ds (Equation 2.11) and is used in equation 2.9
+    ρc_ds = @. (1 - ν) * ρc_ss
+
+    κ_solid =
+        Soil.κ_solid.(ν_ss_om, ν_ss_quartz, p.κ_om, p.κ_quartz, p.κ_minerals)
+    κ_dry = Soil.κ_dry.(ρp, ν, κ_solid, LP.K_therm(earth_param_set))
+    κ_sat_frozen = Soil.κ_sat_frozen.(κ_solid, ν, p.κ_ice)
+    κ_sat_unfrozen = Soil.κ_sat_unfrozen.(κ_solid, ν, p.κ_liq)
+
+    name_map = (;
+        :kersten_number_alpha => :α,
+        :kersten_number_beta => :β,
+        :ice_impedance_omega => :Ω,
+        :temperature_factor_soil_hydraulic_conductivity => :γ,
+        :temperature_reference_soil_hydraulic_conductivity => :γT_ref,
+        :maximum_dry_soil_layer_depth => :d_ds,
+        :soil_momentum_roughness_length => :z_0m,
+        :soil_scalar_roughness_length => :z_0b,
+        :emissivity_bare_soil => :emissivity,
+    )
+    parameters = CP.get_parameter_values(toml_dict, name_map, "Land")
+    PSE = typeof(earth_param_set)
+    FT = CP.float_type(toml_dict)
+    EnergyHydrologyParameters{FT, F, typeof(albedo), C, PSE}(;
+        albedo,
+        ν,
+        ν_ss_om,
+        ν_ss_quartz,
+        ν_ss_gravel,
+        hydrology_cm,
+        K_sat,
+        S_s,
+        θ_r,
+        κ_dry,
+        κ_sat_frozen,
+        κ_sat_unfrozen,
+        ρc_ds,
+        earth_param_set,
+        parameters...,
+        kwargs...,
+    )
+end
+
 """
     EnergyHydrology <: AbstractSoilModel
