+++ src/standalone/Vegetation/radiation.jl
@@ -237,3 +237,121 @@ function canopy_radiant_energy_fluxes!(
         ϵ_canopy * LW_d - 2 * ϵ_canopy * _σ * T_canopy^4 +
         ϵ_canopy * LW_u_ground
 end
+
+## For interfacing with ClimaParams
+
+"""
+    function TwoStreamParameters(FT::AbstractFloat;
+        ld = (_) -> 0.5,
+        α_PAR_leaf = 0.3,
+        τ_PAR_leaf = 0.2,
+        α_NIR_leaf = 0.4,
+        τ_NIR_leaf = 0.25,
+        Ω = 1,
+        n_layers = UInt64(20),
+        kwargs...
+    )
+    function TwoStreamParameters(toml_dict;
+        ld = (_) -> 0.5,
+        α_PAR_leaf = 0.3,
+        τ_PAR_leaf = 0.2,
+        α_NIR_leaf = 0.4,
+        τ_NIR_leaf = 0.25,
+        Ω = 1,
+        n_layers = UInt64(20),
+        kwargs...
+    )
+
+Floating-point and toml dict based constructor supplying default values
+for the TwoStreamParameters struct. Additional parameter values can be directly set via kwargs.
+"""
+TwoStreamParameters(::Type{FT}; kwargs...) where {FT <: AbstractFloat} =
+    TwoStreamParameters(CP.create_toml_dict(FT); kwargs...)
+
+function TwoStreamParameters(
+    toml_dict::CP.AbstractTOMLDict;
+    G_Function = ConstantGFunction(CP.float_type(toml_dict)(0.5)),
+    α_PAR_leaf::F = 0.3,
+    τ_PAR_leaf::F = 0.2,
+    α_NIR_leaf::F = 0.4,
+    τ_NIR_leaf::F = 0.25,
+    Ω = 1,
+    n_layers = UInt64(20),
+    kwargs...,
+) where {F}
+    name_map = (;
+        :wavelength_per_PAR_photon => :λ_γ_PAR,
+        :canopy_emissivity => :ϵ_canopy,
+    )
+
+    parameters = CP.get_parameter_values(toml_dict, name_map, "Land")
+    FT = CP.float_type(toml_dict)
+    # default value for keyword args must be converted manually
+    # automatic conversion not possible to Union types
+    α_PAR_leaf = FT.(α_PAR_leaf)
+    τ_PAR_leaf = FT.(τ_PAR_leaf)
+    α_NIR_leaf = FT.(α_NIR_leaf)
+    τ_NIR_leaf = FT.(τ_NIR_leaf)
+    return TwoStreamParameters{FT, typeof(G_Function), typeof(α_PAR_leaf)}(;
+        G_Function,
+        α_PAR_leaf,
+        τ_PAR_leaf,
+        α_NIR_leaf,
+        τ_NIR_leaf,
+        Ω,
+        n_layers,
+        parameters...,
+        kwargs...,
+    )
+end
+
+"""
+    function BeerLambertParameters(FT::AbstractFloat;
+        ld = (_) -> 0.5,
+        α_PAR_leaf = 0.1,
+        α_NIR_leaf = 0.4,
+        Ω = 1,
+        kwargs...
+    )
+    function BeerLambertParameters(toml_dict;
+        ld = (_) -> 0.5,
+        α_PAR_leaf = 0.1,
+        α_NIR_leaf = 0.4,
+        Ω = 1,
+        kwargs...
+    )
+
+Floating-point and toml dict based constructor supplying default values
+for the BeerLambertParameters struct. Additional parameter values can be directly set via kwargs.
+"""
+BeerLambertParameters(::Type{FT}; kwargs...) where {FT <: AbstractFloat} =
+    BeerLambertParameters(CP.create_toml_dict(FT); kwargs...)
+
+function BeerLambertParameters(
+    toml_dict::CP.AbstractTOMLDict;
+    G_Function = ConstantGFunction(CP.float_type(toml_dict)(0.5)),
+    α_PAR_leaf::F = 0.1,
+    α_NIR_leaf::F = 0.4,
+    Ω = 1,
+    kwargs...,
+) where {F}
+    name_map = (;
+        :wavelength_per_PAR_photon => :λ_γ_PAR,
+        :canopy_emissivity => :ϵ_canopy,
+    )
+
+    parameters = CP.get_parameter_values(toml_dict, name_map, "Land")
+    FT = CP.float_type(toml_dict)
+    # default value for keyword args must be converted manually
+    # automatic conversion not possible to Union types
+    α_PAR_leaf = FT.(α_PAR_leaf)
+    α_NIR_leaf = FT.(α_NIR_leaf)
+    return BeerLambertParameters{FT, typeof(G_Function), typeof(α_PAR_leaf)}(;
+        G_Function,
+        α_PAR_leaf,
+        α_NIR_leaf,
+        Ω,
+        parameters...,
+        kwargs...,
+    )
+end
