+++ test/runtests.jl
@@ -53,6 +53,11 @@ end
     include("standalone/Snow/conservation.jl")
 end
 
+@safetestset "Snow parameter constructors" begin
+    include("standalone/Snow/parameters.jl")
+end
+
+
 # Standalone Soil model tests
 @safetestset "Soil Biogeochemistry module tests" begin
     include("standalone/Soil/Biogeochemistry/biogeochemistry_module.jl")
@@ -84,8 +89,8 @@ end
     include("standalone/Soil/conservation.jl")
 end
 
-@safetestset "Soil spatial parameters" begin
-    include("standalone/Soil/spatial_parameters.jl")
+@safetestset "Soil spatial parameters and parameter constructors" begin
+    include("standalone/Soil/parameters.jl")
 end
 
 # Standalone Surface Water model tests
