+++ test/standalone/Soil/parameters.jl
@@ -63,3 +63,55 @@ const FT = Float64;
     hcm = vg_params.hydrology_cm
     @test eltype(hcm) == ClimaLand.Soil.vanGenuchten{FT}
 end
+
+@testset "Scalar Parameters" begin
+    ν = FT(0.495)
+    K_sat = FT(0.0443 / 3600 / 100) # m/s
+    S_s = FT(1e-3) #inverse meters
+    vg_n = FT(2.0)
+    vg_α = FT(2.6) # inverse meters
+    vg_m = FT(1) - FT(1) / vg_n
+    hcm = ClimaLand.Soil.vanGenuchten{FT}(; α = vg_α, n = vg_n)
+    θ_r = FT(0.1)
+    S_c = hcm.S_c
+
+    ν_ss_om = FT(0.0)
+    ν_ss_quartz = FT(1.0)
+    ν_ss_gravel = FT(0.0)
+    default_params = ClimaLand.Soil.EnergyHydrologyParameters(
+        FT;
+        ν,
+        ν_ss_om,
+        ν_ss_quartz,
+        ν_ss_gravel,
+        hydrology_cm = hcm,
+        K_sat,
+        S_s,
+        θ_r,
+    )
+    @test default_params.emissivity ==
+          LP.get_default_parameter(FT, :emissivity_bare_soil)
+    @test default_params.z_0m ==
+          LP.get_default_parameter(FT, :soil_momentum_roughness_length)
+    @test default_params.z_0b ==
+          LP.get_default_parameter(FT, :soil_scalar_roughness_length)
+
+    overwritten_params = ClimaLand.Soil.EnergyHydrologyParameters(
+        FT;
+        ν,
+        ν_ss_om,
+        ν_ss_quartz,
+        ν_ss_gravel,
+        hydrology_cm = hcm,
+        K_sat,
+        S_s,
+        θ_r,
+        emissivity = FT(0.8),
+        z_0m = FT(0.1),
+        z_0b = FT(0.1),
+    )
+    @test overwritten_params.emissivity == FT(0.8)
+    @test overwritten_params.z_0m == FT(0.1)
+    @test overwritten_params.z_0b == FT(0.1)
+
+end
